+++ src/server/nautical/calib/experiment/mainCalibExperiment.cpp
@@ -9,7 +9,7 @@
 #include <server/nautical/NavCompatibility.h>
 #include <server/common/logging.h>
 #include <server/common/string.h>
-#include <server/nautical/calib/CalibExperiment.h>
+#include <server/nautical/calib/experiment/CalibExperiment.h>
 #include <server/plot/extra.h>
 #include <server/common/Env.h>
 #include <server/common/PathBuilder.h>
